Trying out 7zip

+++ DeploymentScripts/PreparePackage.ps1
@@ -8,7 +8,7 @@ Function Prepare-Package([String] $environmentName, $packageDirectory) {
 
         Expand-Archive -Path $packageDirectory -DestinationPath .\TEMP_PresentationWeb
 
-        Remove-Item -Path $packageDirectory
+        #Remove-Item -Path $packageDirectory
 
         Write-Host "Updating Web.config"
 
@@ -32,8 +32,21 @@ Function Prepare-Package([String] $environmentName, $packageDirectory) {
         $xml.Save($filePathToTask)
 
         Write-Host "Update of Web.config complete"
+
+        $7zipPath = "C:\Program Files\7-Zip\7z.exe"
+
+        if (-not (Test-Path -Path $7zipPath -PathType Leaf)) {
+            throw "7 zip file '$7zipPath' not found"
+        }
+
+        Set-Alias 7zip $7zipPath
+
+        $Source = (Resolve-Path ".\TEMP_PresentationWeb\*")
+        $Target = $packageDirectory
+
+        7zip a -mx=9 $Target $Source
         
-        Compress-Archive -Path (Resolve-Path ".\TEMP_PresentationWeb\*") -CompressionLevel NoCompression -DestinationPath $packageDirectory
+        #Compress-Archive -Path (Resolve-Path ".\TEMP_PresentationWeb\*") -CompressionLevel NoCompression -DestinationPath $packageDirectory
 
         Remove-Item -Path (Resolve-Path ".\TEMP_PresentationWeb") -Recurse -Force
